+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
 ## [1.6.0] - Under development
-- Include SNV pooling options
+- Include SNV pooling options (big update with lots of changes)
+- Include "cov1 = pd.Series(dtype='float64')" in readComparer.py to suppress a future warning
+- Remove cryptic SNVs from the nonredundant_snv_table (https://github.com/MrOlm/inStrain/issues/102)
+- Only produce the run report when run in debug mode
 
 ## [1.5.7] - 2022-02-09
 - Make error log more readable - https://github.com/MrOlm/inStrain/issues/79